# Asset-Manager/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
}

export interface CategoryWithCount {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  quantity: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  image: string;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  quantity: number;
  categoryId: number;
  /** @minLength 1 */
  image: string;
}

export interface ProductListResponse {
  items: Product[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  shipped: "shipped",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  paymentLink?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItem[];
}

export interface CreateOrderItem {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CreateOrderRequest {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @nullable */
  notes?: string | null;
  /** @minItems 1 */
  items: CreateOrderItem[];
}

/**
 * @nullable
 */
export type UpdateOrderRequestStatus =
  | (typeof UpdateOrderRequestStatus)[keyof typeof UpdateOrderRequestStatus]
  | null;

export const UpdateOrderRequestStatus = {
  pending: "pending",
  paid: "paid",
  shipped: "shipped",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderRequest {
  /** @nullable */
  status?: UpdateOrderRequestStatus;
  /** @nullable */
  paymentLink?: string | null;
}

export interface AdminLoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
}

export interface StatusBreakdown {
  status: string;
  count: number;
}

export interface CategoryRevenue {
  categoryId: number;
  categoryName: string;
  revenue: number;
  ordersCount: number;
}

export interface RecentOrderSummary {
  id: number;
  customerName: string;
  totalPrice: number;
  status: string;
  createdAt: string;
}

export interface LowStockProduct {
  id: number;
  name: string;
  quantity: number;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalCategories: number;
  pendingOrders: number;
  ordersByStatus: StatusBreakdown[];
  revenueByCategory: CategoryRevenue[];
  recentOrders: RecentOrderSummary[];
  lowStockProducts: LowStockProduct[];
}

export type ListStorefrontProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
  /**
   * @nullable
   */
  sort?: ListStorefrontProductsSort;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  pageSize?: number | null;
};

export type ListStorefrontProductsSort =
  | (typeof ListStorefrontProductsSort)[keyof typeof ListStorefrontProductsSort]
  | null;

export const ListStorefrontProductsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
} as const;

export type AdminListProductsParams = {
  /**
   * @nullable
   */
  search?: string | null;
};

export type AdminListOrdersParams = {
  /**
   * @nullable
   */
  status?: AdminListOrdersStatus;
};

export type AdminListOrdersStatus =
  | (typeof AdminListOrdersStatus)[keyof typeof AdminListOrdersStatus]
  | null;

export const AdminListOrdersStatus = {
  pending: "pending",
  paid: "paid",
  shipped: "shipped",
  cancelled: "cancelled",
} as const;