# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-commerce store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Admin login (sets session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin-auth]
      summary: Get the currently logged in admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────────────────────── ADMIN CATEGORIES ─────────────────────────
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin-categories]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateCategory
      tags: [admin-categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin-categories]
      summary: Update a category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: adminDeleteCategory
      tags: [admin-categories]
      summary: Delete a category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ───────────────────────── ADMIN PRODUCTS ─────────────────────────
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin-products]
      summary: List all products
      parameters:
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin-products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [admin-products]
      summary: Get a single product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: adminUpdateProduct
      tags: [admin-products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin-products]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ───────────────────────── ADMIN ORDERS ─────────────────────────
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin-orders]
      summary: List all orders
      parameters:
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: ["pending", "paid", "shipped", "cancelled", null]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin-orders]
      summary: Get full details for a single order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: adminUpdateOrder
      tags: [admin-orders]
      summary: Update order status and/or payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ───────────────────────── ADMIN STATS ─────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
      required: [id, name, slug]
...
    CategoryInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
      required: [name, slug]
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        quantity: { type: integer }
        categoryId: { type: integer }
        categoryName:
          type: ["string", "null"]
        image: { type: string }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - name
        - description
        - price
        - quantity
        - categoryId
        - image
        - createdAt
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        quantity: { type: integer, minimum: 0 }
        categoryId: { type: integer }
        image: { type: string, minLength: 1 }
      required: [name, description, price, quantity, categoryId, image]
...
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage:
          type: ["string", "null"]
        quantity: { type: integer }
        price: { type: number }
      required: [id, productId, productName, quantity, price]
    Order:
      type: object
      properties:
        id: { type: integer }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        totalPrice: { type: number }
        status:
          type: string
          enum: [pending, paid, shipped, cancelled]
        paymentLink:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - customerName
        - phone
        - address
        - totalPrice
        - status
        - createdAt
        - items
...
    UpdateOrderRequest:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: [pending, paid, shipped, cancelled, null]
        paymentLink:
          type: ["string", "null"]
    AdminLoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
      required: [id, email]
...