# Asset-Manager/lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  quantity: number;
  categoryId: number;
  /** @minLength 1 */
  image: string;
}