# Asset-Manager/lib/api-zod/src/generated/types/updateOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderRequestStatus } from "./updateOrderRequestStatus";

export interface UpdateOrderRequest {
  /** @nullable */
  status?: UpdateOrderRequestStatus;
  /** @nullable */
  paymentLink?: string | null;
}